+++ packages/terra-hyperlink/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Fixed
+  * Fixed focus box bug for hyperlink.
+
 ## 2.51.4 - (May 3, 2022)
 
 * Changed